# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api SDE for mobile sde",
        "description": "Implementation of Swagger with in Laravel",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://haki44.eu/api",
            "description": "API SDE"
        }
    ],
    "paths": {
...
        "/adventure-booking/{id}": {
            "get": {
                "tags": [
                    "Adventure Booking"
                ],
                "summary": "Show one adventure booking",
                "description": "Show one adventure booking",
                "operationId": "ShowAdventureBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of adventure booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Adventure Booking"
                ],
                "summary": "Update one adventure booking",
                "description": "Update one adventure booking",
                "operationId": "UpdateAdventureBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of adventure booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/adventure": {
            "get": {
                "tags": [
                    "Adventures"
                ],
                "summary": "Get List Of adventures",
                "description": "Returns all adventures",
                "operationId": "indexAdventure",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Adventures"
                ],
                "summary": "Store adventure",
                "description": "Store adventure",
                "operationId": "storeAdventure",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/adventure/{id}": {
            "get": {
                "tags": [
                    "Adventures"
                ],
                "summary": "Show one adventure",
                "description": "Show one adventure",
                "operationId": "ShowAdventure",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of adventure",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Adventures"
                ],
                "summary": "Update one adventure",
                "description": "Update one adventure",
                "operationId": "updateAdventure",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of adventure",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Adventures"
                ],
                "summary": "Delete one adventure",
                "description": "Delete one adventure",
                "operationId": "deleteAdventure",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of adventure",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}